refactor: modularize navigation helpers
